+++ Scalexpert/Plugin/Block/Payment/Info.php
@@ -1,4 +1,12 @@
 <?php
+/**
+ * Copyright © Scalexpert.
+ * This file is part of Scalexpert plugin for Magento 2. See COPYING.md for license details.
+ *
+ * @author    Scalexpert (https://scalexpert.societegenerale.com/)
+ * @copyright Scalexpert
+ * @license   https://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ */
 namespace Scalexpert\Plugin\Block\Payment;
 
 use Magento\Framework\Serialize\Serializer\Json;
@@ -48,7 +56,8 @@ class Info extends \Magento\Payment\Block\Info
             $info['buyer_financedAmount'] = isset($add['buyer_financedAmount']) ? $add['buyer_financedAmount'] : false;
             $info['consolidated_status'] = isset($add['consolidated_status']) ? $add['consolidated_status'] : false;
             $info['method_title'] = $payment->getMethodInstance()->getTitle();
+            $info['status_message_error'] = isset($add['status_message_error']) ? $add['status_message_error'] : false;
         }
-            return $info;
+        return $info;
     }
 }